Marianne's stuff in the ui section removed payment model updated PaymentLogic updated -> swapped to two methods Further work on paymentDAL edit on henry's table popup edit on login form to reflect removal of marianne's pieces

+++ ProjectChapeau/ChapeauModel/OrderItems.cs
@@ -10,10 +10,10 @@ namespace Chapeau_Model
     {
         public string ItemName;
         public int Quantity;
-        public double ItemPrice;
-        public double Vat;
+        public float ItemPrice;
+        public float Vat;
 
-        public OrderItems(string itemName, int quantity, double itemPrice, double vat)
+        public OrderItems(string itemName, int quantity, float itemPrice, float vat)
         {
             ItemName = itemName;
             Quantity = quantity;
